+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,78 +1,13 @@
 package roomescape.repository;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
 
-@Repository
-public class ReservationRepository {
+public interface ReservationRepository {
 
-    private static final String TABLE_NAME = "reservation";
+    List<Reservation> findAll();
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert simpleJdbcInsert;
+    Reservation save(Reservation reservation);
 
-    private final RowMapper<Reservation> reservationRowMapper = (resultSet, rowNum) -> {
-        Reservation reservation
-                = new Reservation(
-                resultSet.getLong("id"),
-                resultSet.getString("name"),
-                LocalDate.parse(resultSet.getString("date")),
-                new ReservationTime(
-                        resultSet.getLong("time_id"),
-                        LocalTime.parse(resultSet.getString("start_at"))
-                )
-        );
-        return reservation;
-    };
-
-    public ReservationRepository(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
-                .withTableName(TABLE_NAME)
-                .usingGeneratedKeyColumns("id");
-    }
-
-    public List<Reservation> findAll() {
-        String sql = """
-                SELECT
-                    r.id as reservation_id,
-                    r.name,
-                    r.date,
-                    t.id as time_id,
-                    t.start_at as time_value
-                FROM reservation as r
-                inner join reservation_time as t
-                on r.time_id = t.id
-                """;
-        return jdbcTemplate.query(sql, reservationRowMapper);
-    }
-
-    public Reservation save(Reservation reservation) {
-        SqlParameterSource parameterSource = new MapSqlParameterSource()
-                .addValue("name", reservation.getName())
-                .addValue("date", reservation.getDate())
-                .addValue("time_id", reservation.getReservationTime().getId());
-        Long id = simpleJdbcInsert.executeAndReturnKey(parameterSource).longValue();
-        return new Reservation(
-                id,
-                reservation.getName(),
-                reservation.getDate(),
-                reservation.getReservationTime()
-        );
-    }
-
-    public int deleteById(Long id) {
-        String sql = "DELETE FROM " + TABLE_NAME + " WHERE id = ?";
-        return jdbcTemplate.update(sql, id);
-    }
+    int deleteById(Long id);
 }
