+++ src/main/java/roomescape/repository/ReservationTimeJdbcRepository.java
@@ -1,0 +1,59 @@
+package roomescape.repository;
+
+import java.time.LocalTime;
+import java.util.List;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.domain.ReservationTime;
+
+@Repository
+public class ReservationTimeJdbcRepository implements ReservationTimeRepository{
+
+    private static final String TABLE_NAME = "reservation_time";
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert simpleJdbcInsert;
+
+    private final RowMapper<ReservationTime> reservationTimeRowMapper = (resultSet, rowNum) -> {
+        ReservationTime reservationTime
+                = new ReservationTime(
+                resultSet.getLong("id"),
+                LocalTime.parse(resultSet.getString("start_at"))
+        );
+        return reservationTime;
+    };
+
+    public ReservationTimeJdbcRepository(JdbcTemplate jdbcTemplate, DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName(TABLE_NAME)
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT id, start_at FROM " + TABLE_NAME;
+        return jdbcTemplate.query(sql, reservationTimeRowMapper);
+    }
+
+    public ReservationTime findByTimeId(Long timeId) {
+        String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, reservationTimeRowMapper, timeId);
+    }
+
+    public ReservationTime save(ReservationTime reservationTime) {
+        SqlParameterSource parameterSource = new MapSqlParameterSource()
+                .addValue("start_at", reservationTime.getStartAt());
+        Long id = simpleJdbcInsert.executeAndReturnKey(parameterSource).longValue();
+        return new ReservationTime(id, reservationTime.getStartAt());
+    }
+
+    public int deleteById(Long id) {
+        String sql = "DELETE FROM " + TABLE_NAME + " WHERE id = ?";
+        return jdbcTemplate.update(sql, id);
+    }
+}
